+++ app/components/MintNBTC.tsx
@@ -1,55 +1,26 @@
 import { Card, CardContent } from "./ui/card";
-import { BitcoinBalance } from "./BitcoinBalance";
-import { Link } from "@remix-run/react";
 import { Button } from "./ui/button";
-import { FormProvider, useForm } from "react-hook-form";
 import { FormInput } from "./form/FormInput";
+import { FormProvider, useForm } from "react-hook-form";
+import { useContext, useCallback } from "react";
+import { WalletContext } from "~/providers/ByieldWalletProvider";
+import { ByieldWallet } from "~/types";
+import { SuiModal } from "./Wallet/SuiWallet/SuiModal";
+import { useSignAndExecuteTransaction, useSuiClient } from "@mysten/dapp-kit";
+import { Transaction } from "@mysten/sui/transactions";
+import { useNetworkVariables } from "~/networkConfig";
+import { suiToMist } from "~/util/util";
+import { useToast } from "~/hooks/use-toast";
 
-const PERCENTAGE = [
-	{
-		id: "percentage-1",
-		value: 25,
-	},
-	{
-		id: "percentage-2",
-		value: 50,
-	},
-	{
-		id: "percentage-3",
-		value: 75,
-	},
-	{
-		id: "percentage-4",
-		value: 100,
-	},
-];
-
-function Percentage() {
-	return (
-		<div className="flex justify-between mb-4 space-x-2">
-			{PERCENTAGE.map(({ id, value }) => (
-				<Button key={id} variant="outline" className="bg-azure-10 w-full flex-1 flex">
-					{value}%
-				</Button>
-			))}
-		</div>
-	);
-}
-
-interface ExchangeRateProps {
-	fee: number;
+interface FeeProps {
 	youReceive: number;
 }
 
-function Fee({ fee, youReceive }: ExchangeRateProps) {
+function Fee({ youReceive }: FeeProps) {
 	return (
-		<Card className="p-4 bg-azure-10 rounded-2xl h-24">
-			<CardContent className="flex flex-col justify-between h-full p-0">
-				<div className="flex justify-between">
-					<p className="text-gray-400">Fixed Fee</p>
-					<p>{fee} Satoshi</p>
-				</div>
-				<div className="flex justify-between">
+		<Card className="p-4 bg-azure-10 rounded-2xl h-14">
+			<CardContent className="flex flex-col justify-between p-0">
+				<div className="flex justify-between gap-2">
 					<p className="text-gray-400">You Receive</p>
 					<p>{youReceive} nBTC</p>
 				</div>
@@ -58,70 +29,100 @@ function Fee({ fee, youReceive }: ExchangeRateProps) {
 	);
 }
 
-interface MinBTCForm {
-	numberOfBTC: number;
-	suiAddress: string;
+interface OtcBuyForm {
+	suiAmount: number;
 }
 
-interface MintBTCProps {
-	availableBalance: number;
-	suiAddress: string;
-}
-
-export function MintNBTC({ availableBalance, suiAddress }: MintBTCProps) {
-	const mintBTCForm = useForm<MinBTCForm>({
-		defaultValues: {
-			// TODO: make it dynamic
-			numberOfBTC: 1,
-			suiAddress,
-		},
+export function BuyNBTC() {
+	const { toast } = useToast();
+	const { connectedWallet } = useContext(WalletContext);
+	const client = useSuiClient();
+	const { nbtcOTC } = useNetworkVariables();
+	const { packageId, module, swapFunction, vaultId, pricePerNBTCInSUI } = nbtcOTC;
+	const isSuiWalletConnected = connectedWallet === ByieldWallet.SuiWallet;
+	const { mutate: signAndExecuteTransaction } = useSignAndExecuteTransaction({
+		execute: async ({ bytes, signature }) =>
+			await client.executeTransactionBlock({
+				transactionBlock: bytes,
+				signature,
+				options: {
+					showRawEffects: true,
+					showObjectChanges: true,
+				},
+			}),
 	});
-	const { handleSubmit, watch } = mintBTCForm;
-	const numberOfBTC = watch("numberOfBTC");
 
-	// santoshi. 1BTC = 10^8 satoshi
-	const fee = 0.00000001;
+	const otcBuyForm = useForm<OtcBuyForm>({
+		mode: "all",
+		reValidateMode: "onChange",
+	});
+	const { watch } = otcBuyForm;
+	const suiAmount = watch("suiAmount");
+	const amountOfNBTC = suiAmount / pricePerNBTCInSUI;
+	const isAmountOfNBTCValid = amountOfNBTC > 0;
+
+	const handleTransaction = useCallback(
+		({ suiAmount }: OtcBuyForm) => {
+			const suiAmountMist = suiToMist(suiAmount);
+			const transaction = new Transaction();
+			const [coins] = transaction.splitCoins(transaction.gas, [transaction.pure.u64(suiAmountMist)]);
+			transaction.moveCall({
+				target: `${packageId}::${module}::${swapFunction}`,
+				arguments: [
+					transaction.object(vaultId), // Vault object
+					coins, // Coin<SUI> argument
+				],
+			});
+
+			signAndExecuteTransaction(
+				{
+					transaction,
+				},
+				{
+					onSuccess: () => {
+						toast({
+							title: "Buy nBTC",
+							description: `Transaction succeeded`,
+						});
+					},
+					onError: (error) => {
+						toast({
+							title: "Buy nBTC",
+							description: `Transaction failed: ${error.message}`,
+							variant: "destructive",
+						});
+					},
+				},
+			);
+		},
+		[module, packageId, signAndExecuteTransaction, swapFunction, toast, vaultId],
+	);
 
 	return (
-		<FormProvider {...mintBTCForm}>
-			<form
-				onSubmit={handleSubmit((formData) => {
-					// TODO: handle the nbtc form data
-					console.log("Depositing with Sui Address:", formData);
-				})}
-			>
+		<FormProvider {...otcBuyForm}>
+			<form onSubmit={otcBuyForm.handleSubmit(handleTransaction)} className="w-1/2">
 				<Card>
 					<CardContent className="p-6 rounded-lg text-white flex flex-col gap-4 bg-azure-10">
-						<BitcoinBalance availableBalance={availableBalance} />
 						<FormInput
 							required
-							name="numberOfBTC"
+							name="suiAmount"
 							type="number"
-							placeholder="number"
-							rightAdornments={<span className="text-sm font-medium w-20">~$0 USD</span>}
+							placeholder="Enter SUI amount"
 							className="h-16"
+							rules={{
+								validate: {
+									positive: (value: string) =>
+										Number(value) > 0 || "SUI amount must be greater than 0",
+								},
+							}}
 						/>
-						<Percentage />
-						<FormInput
-							required
-							name="suiAddress"
-							type="text"
-							placeholder="Enter Your Sui Address..."
-							value={suiAddress}
-							className="h-16"
-						/>
-						<Fee fee={10} youReceive={numberOfBTC - fee} />
-						<Button type="submit">Deposit BTC and mint nBTC</Button>
-						<div className="flex justify-between">
-							<span>TX ID: b99d9a361ac9db3...</span>
-							<Link
-								target="_blank"
-								to={"https://v3.tailwindcss.com/docs/font-weight"}
-								rel="noreferrer"
-							>
-								Track confirmation in explorer
-							</Link>
-						</div>
+						{isAmountOfNBTCValid && <Fee youReceive={amountOfNBTC} />}
+						{isAmountOfNBTCValid && (
+							<span className="tracking-tighter text-gray-500 text-sm dark:text-gray-400">
+								This is a fixed price buy. The price is 25,000 SUI / BTC.
+							</span>
+						)}
+						{isSuiWalletConnected ? <Button type="submit">Buy</Button> : <SuiModal />}
 					</CardContent>
 				</Card>
 			</form>
